+++ beymax/bots/game_systems/base.py
