+++ beymax/bots/game_systems/poker/__init__.py
@@ -3,7 +3,7 @@ import asyncio
 import random
 import json
 import os
-from ...utils import DBView, getname
+from ....utils import DBView, getname
 from ..base import GameEndException, PhasedGame
 from .utils import Card, Hand, Deck, PokerError
 from .texas_hold_em import __GAME_DEF as texas, __RESTORE as texas_restore
