+++ beymax/bots/game_systems/poker/blackjack.py
@@ -1,5 +1,5 @@
 import asyncio
-from ...utils import DBView, getname
+from ....utils import DBView, getname
 from ..base import GameEndException
 from .utils import FreePhase, LockedPhase, Card, Hand, strike_if
