+++ beymax/bots/game_systems/poker/texas_hold_em.py
@@ -1,5 +1,5 @@
 import asyncio
-from ...utils import DBView, getname
+from ....utils import DBView, getname
 from .utils import FreePhase, LockedPhase, Hand, strike_if
 
 # For handling side-pots:
