+++ beymax/bots/poll.py
@@ -1,0 +1,72 @@
+from ..core import CommandSuite
+from ..utils import getname
+from ..args import Arg
+import discord
+import asyncio
+
+Polls = CommandSuite('Polls')
+
+@Polls.subscribe('before:firstready')
+async def fixme_add_polls(self, event):
+    # This is a dumb pattern. CommandSuites should retain their own state
+    self.polls = {}
+
+@Polls.add_command(
+    'poll',
+    Arg('title', help="Poll Title"),
+    Arg("options", nargs='*', help="Poll options"),
+    delimiter='|'
+)
+async def cmd_poll(self, message, title, options):
+    """
+    `$!poll <poll title> | [Option 1] | [Option 2] | [etc...]` : Creates a poll
+    Example: `$!poll Is $NAME cool? | Yes | Definitely`
+    """
+    #The argparse API is killing the blank handling, but I think that's okay
+    opts = [
+        (opt.rstrip() if '~<blank>' not in opt else opt)
+        for opt in options
+    ]
+    if sum(1 for opt in opts if not len(opt)):
+        await self.send_message(
+            message.author,
+            "Your poll command contained trailing or adjacent `|` characters"
+            " which resulted in blank fields that I'm going to ignore. If"
+            " the blank fields were intentional, add `~<blank>` into each"
+            " field that you want to leave blank"
+        )
+    opts = [opt.replace('~<blank>', '') for opt in opts if len(opt)]
+    body = getname(message.author)+" has started a poll:\n"
+    body+=title+"\n"
+    body+="\n".join((
+            "%d) %s"%(num+1, opt)
+            for (num, opt) in
+            enumerate(opts)
+        ))
+    body+="\n\nReact with your vote"
+    target = await self.send_message(
+        message.channel,
+        body,
+        skip_debounce=True
+    )
+    for i in range(1,len(opts)+1):
+        await target.add_reaction(
+            (b'%d\xe2\x83\xa3'%i).decode()#hack to create number emoji reactions
+        )
+    if not isinstance(message.channel, discord.abc.PrivateChannel):
+        try:
+            await message.delete()
+        except:
+            print("Warning: Unable to delete poll source message")
+        self.polls[target.id] = (message.author, {self.user.id})
+
+@Polls.subscribe('reaction_add')
+async def on_reaction_add(self, event, reaction, user):
+    if reaction.message.id in self.polls:
+        creator, reactors = self.polls[reaction.message.id]
+        if user.id not in reactors:
+            await self.send_message(
+                creator,
+                getname(user)+" has voted on your poll in "+reaction.message.channel.name
+            )
+            self.polls[reaction.message.id][1].add(user.id)
